:recycle: One header less
